+++ bpxe/src/activity/script_task.rs
@@ -74,7 +74,7 @@ impl FlowNode for Task {
         }
     }
 
-    fn get_state(&self) -> flow_node::State {
+    fn get_state(&mut self) -> flow_node::State {
         flow_node::State::ScriptTask(self.state.clone())
     }