+++ bpxe/src/event/intermediate_throw_event.rs
@@ -50,7 +50,7 @@ impl FlowNode for IntermediateThrowEvent {
         }
     }
 
-    fn get_state(&self) -> flow_node::State {
+    fn get_state(&mut self) -> flow_node::State {
         flow_node::State::IntermediateThrowEvent(self.state.clone())
     }