+++ bpxe/src/event/start_event.rs
@@ -63,7 +63,7 @@ impl FlowNode for StartEvent {
         }
     }
 
-    fn get_state(&self) -> flow_node::State {
+    fn get_state(&mut self) -> flow_node::State {
         flow_node::State::StartEvent(self.state.clone())
     }