+++ bpxe/src/gateway/event_based.rs
@@ -53,7 +53,7 @@ impl FlowNode for Gateway {
         self.process = Some(process);
     }
 
-    fn get_state(&self) -> flow_node::State {
+    fn get_state(&mut self) -> flow_node::State {
         flow_node::State::EventBasedGateway(self.state.clone())
     }