+++ bpxe/src/language/rhai.rs
@@ -4,6 +4,8 @@
 use super::{Engine, EngineInfo, EvaluationError};
 use crate::bpmn::schema::{FormalExpression, Script, ScriptTask};
 use async_trait::async_trait;
+use num_bigint::BigInt;
+use std::any::{Any, TypeId};
 use std::ops::Deref;
 use std::sync::Arc;
 use tokio::task;
@@ -29,12 +31,50 @@ impl Deref for Rhai {
     }
 }
 
+// Translate types that Rhai doesn't support
+//
+// This is necessary to establish types that are shared between different languages
+// so that the engine doesn't need to be aware of the language being used.
+macro_rules! type_translation {
+    ($expr: expr, $t: ty, $t1: ty, $ident: ident, $convert: expr) => {
+        if TypeId::of::<T>() == TypeId::of::<$t1>() {
+            let result: Result<$t, _> = $expr;
+            return result.map(|$ident| {
+                let v1 = $convert;
+                let any: &dyn Any = &v1;
+                any.downcast_ref::<T>().unwrap().clone()
+            });
+        }
+    };
+    ($expr: expr) => {{
+        type_translation!($expr, i64, usize, value, value as usize);
+        type_translation!($expr, i64, BigInt, value, BigInt::from(value));
+        $expr
+    }};
+}
+
 impl Rhai {
     /// Tries to get a mutable reference to the engine
     ///
     /// Primarily useful for custom setup, tests, etc.
     pub fn engine_mut(&mut self) -> Option<&mut rhai::Engine> {
         Arc::get_mut(&mut self.engine)
+    }
+
+    #[inline]
+    fn internal_eval<T>(&self, expr: &str) -> Result<T, Box<rhai::EvalAltResult>>
+    where
+        T: Clone + Send + Sync + 'static,
+    {
+        type_translation!(self.engine.eval(expr))
+    }
+
+    #[inline]
+    fn internal_eval_expression<T>(&self, expr: &str) -> Result<T, Box<rhai::EvalAltResult>>
+    where
+        T: Clone + Send + Sync + 'static,
+    {
+        type_translation!(self.engine.eval_expression(expr))
     }
 }
 
@@ -81,8 +121,8 @@ impl Engine<FormalExpression> for Rhai {
             None => return Err(EvaluationError::Empty),
             Some(ref content) => {
                 let expr = content.to_owned();
-                let engine = self.engine.clone();
-                Ok(task::spawn_blocking(move || engine.eval_expression(&expr)).await??)
+                let engine = self.clone();
+                Ok(task::spawn_blocking(move || engine.internal_eval_expression(&expr)).await??)
             }
         }
     }
@@ -100,8 +140,8 @@ impl Engine<ScriptTask> for Rhai {
                 content: Some(ref content),
             }) => {
                 let expr = content.to_owned();
-                let engine = self.engine.to_owned();
-                Ok(task::spawn_blocking(move || engine.eval(&expr)).await??)
+                let engine = self.clone();
+                Ok(task::spawn_blocking(move || engine.internal_eval(&expr)).await??)
             }
         }
     }
@@ -144,5 +184,67 @@ mod tests {
             })
             .await
             .unwrap());
+    }
+
+    #[tokio::test]
+    async fn rhai_bigint_support() {
+        use num_bigint::BigInt;
+        let e = Rhai::new();
+        assert_eq!(
+            e.eval::<BigInt>(&ScriptTask {
+                script: Some(Script {
+                    content: Some("100".into())
+                }),
+                ..Default::default()
+            })
+            .await
+            .unwrap(),
+            BigInt::from(100)
+        );
+    }
+
+    #[tokio::test]
+    async fn rhai_bigint_support_expr() {
+        use num_bigint::BigInt;
+        let e = Rhai::new();
+        assert_eq!(
+            e.eval::<BigInt>(&FormalExpression {
+                content: Some("100".into()),
+                ..Default::default()
+            })
+            .await
+            .unwrap(),
+            BigInt::from(100)
+        );
+    }
+
+    #[tokio::test]
+    async fn rhai_usize_support() {
+        let e = Rhai::new();
+        assert_eq!(
+            e.eval::<usize>(&ScriptTask {
+                script: Some(Script {
+                    content: Some("100".into())
+                }),
+                ..Default::default()
+            })
+            .await
+            .unwrap(),
+            100
+        );
+    }
+
+    #[tokio::test]
+    async fn rhai_usize_support_expr() {
+        let e = Rhai::new();
+        assert_eq!(
+            e.eval::<usize>(&FormalExpression {
+                content: Some("100".into()),
+                ..Default::default()
+            })
+            .await
+            .unwrap(),
+            100
+        );
     }
 }