+++ bpxe/src/lib.rs
@@ -14,6 +14,7 @@
 //! resumed with little to no consideration when a failure happen.
 pub mod activity;
 pub mod bpmn;
+pub mod data_object;
 pub mod event;
 pub mod flow_node;
 pub mod gateway;