+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Inclusive gateway support
 - `scriptTask` activity support
 - `standardLoopCharacteristics` support for activities
+- `multiInstanceLoopCharacteristics` support for activities
+- Basic support for data objects
 
 ### Fixed